requirements.txt
# Python 3.11.x Required for optimal LLM performance and compatibility
# Ensure you're using Python 3.11.x before installing dependencies

# Python 3.11.x Required for optimal LLM performance and compatibility
# Ensure you're using Python 3.11.x before installing dependencies

click>=8.1.8
# Keep crewai at version 0.130.0
# Other package versions are flexible and can be changed
# Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means. But we are expecting you to use the crewai version 0.130.0.
crewai==0.130.0 
crewai-tools==0.47.1
fastapi==0.110.3
google-api-core>=2.11.0
Jinja2==3.1.4
jsonschema==4.22.0
## langchain-core not used; removed to avoid conflicts
## langsmith will be resolved transitively by crewai-tools/embedchain
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai>=1.68.2
tiktoken>=0.8.0
# Align OpenTelemetry versions with crewai>=0.130.0 requirements
opentelemetry-api>=1.30.0
opentelemetry-exporter-otlp-proto-common>=1.30.0
opentelemetry-exporter-otlp-proto-grpc>=1.30.0
opentelemetry-exporter-otlp-proto-http>=1.30.0
opentelemetry-instrumentation>=0.51b0
opentelemetry-instrumentation-asgi>=0.51b0
opentelemetry-instrumentation-fastapi>=0.51b0
opentelemetry-proto>=1.30.0
opentelemetry-sdk>=1.30.0
opentelemetry-semantic-conventions>=0.51b0
opentelemetry-util-http>=0.51b0
pandas==2.2.2
pillow==10.3.0
pip==24.0
protobuf>=5.0.0,<6.0.0
pydantic>=2.4.2
# Allow resolver to pick compatible pydantic-core for pydantic v2
# ======================================================================

python-multipart==0.0.20
# Authentication and Security Dependencies
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
# Database Dependencies
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
# Redis for rate limiting and caching
aioredis==2.0.1
# Additional security
bleach==6.1.0
# Email validation
email-validator==2.1.0